+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY1WZR737NGPGZK5N3FBR08M"
+        "01KY2EENPRTG67R64GBQSFHR91"
     )
     r"""The unique identifier of the data source"""