+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-07-21T08:34:12.428Z"
+    created: Optional[str] = "2026-07-21T13:39:04.321Z"
 
-    updated: Optional[str] = "2026-07-21T08:34:12.428Z"
+    updated: Optional[str] = "2026-07-21T13:39:04.321Z"
 
     updated_by_id: OptionalNullable[str] = UNSET