+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY1WZR71GB6VMXZCPQPP23H1"
+        "01KY2EENPNQH7WZBWDHZ1PBDA3"
     )
     r"""The unique identifier of the data source"""